Consolidate jitdump functionality into JitDump.cpp
All functions related to jitdump generation are placed into JitDump.cpp
and only the functions required external linkage are placed into the
corresponding header file. This is the first step towards refactoring
and improving the jitdump functionality and documentation.
